:truck: Move spans tab search section into its own file
